Load_Neo4j/requirements.txt
neo4j==5.12.0
python-dotenv==1.0.0